Generate component state when delegate state gets updated
COAND-935
